fix: Correct usage of twiml response
In handleAgentConnected we were responding with a twiml to pause for 5
seconds to give the system a chance to process, then we sent a call
modification. This was an unnecessary step, we now send teh correct
twiml response the first time and don't modify the call afterwards.

This commit resolves #24